+++ .src/array_iterator.inl
@@ -88,7 +88,7 @@ namespace jluna
     Array<V, R>::ConstIterator::operator T() const
     {
         static jl_function_t* getindex = jl_get_function(jl_base_module, "getindex");
-        return unbox<T>(jluna::safe_call(getindex, _owner->operator jl_value_t *(), box<size_t>(_index + 1)));
+        return unbox<T>(jluna::safe_call(getindex, _owner->operator jl_value_t *(), box<uint64_t>(_index + 1)));
     }
 
     template<is_boxable V, size_t R>
@@ -106,7 +106,7 @@ namespace jluna
         static jl_function_t* setindex = jl_get_function(jl_base_module, "setindex!");
 
         gc_pause;
-        jl_call3(setindex, _owner->operator jl_value_t *(), box(value), box(_index + 1));
+        jl_call3(setindex, _owner->operator jl_value_t *(), box(value), box((uint64_t)(_index + 1)));
         gc_unpause;
 
         return *this;
@@ -117,6 +117,6 @@ namespace jluna
     Array<V, R>::Iterator::operator T() const
     {
         static jl_function_t* getindex = jl_get_function(jl_base_module, "getindex");
-        return unbox<T>(jluna::safe_call(getindex, _owner->operator jl_value_t *(), box<size_t>(_index + 1)));
+        return unbox<T>(jluna::safe_call(getindex, _owner->operator jl_value_t *(), box<uint64_t>(_index + 1)));
     }
 }
